use function iniGetInt64CorrectValue

+++ src/storage/storage_config.h
@@ -84,7 +84,7 @@ typedef struct {
     int io_depth_per_read_thread;
     double reserved_space_per_disk;
     int max_trunk_files_per_subdir;
-    int64_t trunk_file_size;
+    uint32_t trunk_file_size;
     int discard_remain_space_size;
     int trunk_prealloc_threads;
     int fd_cache_capacity_per_read_thread;
